requirements.txt
certifi==2016.2.28
fire==0.1.3
lmdb==0.94
numpy==1.14.2
Pillow==5.3.0
pyarrow==0.10.0
six==1.11.0
tqdm==4.26.0
u-msgpack-python==2.5.0
